refactor: move AttributeInput out of rmrk
