+++ components/create/CreateNft.vue
@@ -305,7 +305,7 @@ import CreateNftPreview from './CreateNftPreview.vue'
 import type { ActionMintToken, ActionList, TokenToList } from '@/composables/transaction/types'
 import ChooseCollectionDropdown from '@/components/common/ChooseCollectionDropdown.vue'
 import BasicSwitch from '@/components/shared/form/BasicSwitch.vue'
-import CustomAttributeInput from '@/components/rmrk/Create/CustomAttributeInput.vue'
+import CustomAttributeInput from '@/components/create/CustomAttributeInput.vue'
 import RoyaltyForm from '@/components/bsx/Create/RoyaltyForm.vue'
 import MintConfirmModal from '@/components/create/Confirm/MintConfirmModal.vue'
 import resolveQueryPath from '@/utils/queryPathResolver'
